New chattee and new checkpoints

+++ app/src/main/java/com/hippovio/child/database/local/entities/MessageReadCheckpoint.java
@@ -36,7 +36,7 @@ public class MessageReadCheckpoint {
 
     @ColumnInfo(name = "chatee_id")
     @NonNull
-    private int chateeId;
+    private Long chateeId;
 
     @TypeConverters({SourceConverter.class})
     @ColumnInfo(name = "message_source")
